added emerald apple, renamed package
