+++ src/main/java/io/github/lukas2005/multicraft/items/itemclasses/simple/ItemLlamaFur.java
@@ -1,4 +1,4 @@
-package io.github.lukas2005.multicraft.items.itemclasses.basematerials;
+package io.github.lukas2005.multicraft.items.itemclasses.simple;
 
 import io.github.lukas2005.multicraft.items.ItemBase;
 import net.minecraft.creativetab.CreativeTabs;
